requirements.txt
# Core
apache-airflow==2.8.0
psycopg2-binary==2.9.9
SQLAlchemy==1.4.51          # Airflow 2.8 requires <2.0

# Scraping
selenium==4.18.1
undetected-chromedriver==3.5.5
selenium-stealth==1.0.6
fake-useragent==1.4.0

# Data
pandas==2.2.1

# Utils
python-dotenv==1.0.1
